requirements.txt
requests
nltk
openai
python-docx
